+++ src/main/java/pl/mmo/services/KlaserServiceJPAImpl.java
@@ -54,9 +54,10 @@ public class KlaserServiceJPAImpl implements KlaserService {
 
     @Override
     public Optional<Boolean> deleteById(Long id) {
-        bazaDanych.delete(id.intValue());
+        bazaDanych.delete(id);
         return Optional.of(Boolean.TRUE);
     }
+
 
     @Override
     public List<Film> findLatest3() {
